refactoring admin console server code
