+++ website/test/admin.html
@@ -101,6 +101,18 @@
                 <span id="dataMatch">-</span>
               </div>
             </div>
+            <div class="tool-grid" style="margin-bottom: 10px">
+              <button
+                id="forceReloadLive"
+                class="tool-btn destructive"
+                disabled
+              >
+                Force DB → Live
+              </button>
+              <button id="forceSaveLive" class="tool-btn" disabled>
+                Force Save DB ← Live
+              </button>
+            </div>
             <div class="shared-data-grid">
               <div class="shared-section">
                 <h4>Direct Method Data (Admin Console)</h4>
